Auth to endpoints

+++ routers/movies.py
@@ -14,14 +14,20 @@ router = APIRouter(
 
 @router.get("/", status_code=status.HTTP_200_OK)
 async def read_all(
+    user: user_dependency,
     db: db_dependency,
 ):
+    if user["role"] not in ["admin", "user"]:
+        raise HTTPException(
+            status_code=403,
+            detail="Permission denied. Only users have access to this operation.",
+        )
     return db.query(DBMovie).all()
 
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
 async def create(user: user_dependency, db: db_dependency, movie_request: MovieRequest):
-    if user.role != "admin" or user is None:
+    if user["role"] != "admin" or user is None:
         raise HTTPException(
             status_code=403,
             detail="Permission denied. Only admin users have access to this operation.",
@@ -32,7 +38,12 @@ async def create(user: user_dependency, db: db_dependency, movie_request: MovieR
 
 
 @router.get("/{movie_id}", status_code=status.HTTP_200_OK)
-async def read(db: db_dependency, movie_id: int = Path(gt=0)):
+async def read(user: user_dependency, db: db_dependency, movie_id: int = Path(gt=0)):
+    if user["role"] not in ["admin", "user"]:
+        raise HTTPException(
+            status_code=403,
+            detail="Permission denied. Only users have access to this operation.",
+        )
     movie_model = db.query(DBMovie).filter(DBMovie.id == movie_id).first()
     if movie_model is None:
         raise HTTPException(status_code=404, detail="Movie not found")
@@ -41,8 +52,16 @@ async def read(db: db_dependency, movie_id: int = Path(gt=0)):
 
 @router.put("/{movie_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def update(
-    db: db_dependency, movie_request: MovieRequest, movie_id: int = Path(gt=0)
+    user: user_dependency,
+    db: db_dependency,
+    movie_request: MovieRequest,
+    movie_id: int = Path(gt=0),
 ):
+    if user["role"] != "admin" or user is None:
+        raise HTTPException(
+            status_code=403,
+            detail="Permission denied. Only admin users have access to this operation.",
+        )
     movie_model = db.query(DBMovie).filter(DBMovie.id == movie_id).first()
     if movie_model is None:
         raise HTTPException(status_code=404, detail="Movie not found")
@@ -55,8 +74,13 @@ async def update(
 
 
 @router.delete("/{movie_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete(db: db_dependency, movie_id: int = Path(gt=0)):
+async def delete(user: user_dependency, db: db_dependency, movie_id: int = Path(gt=0)):
     movie_model = db.query(DBMovie).filter(DBMovie.id == movie_id).first()
+    if user["role"] != "admin" or user is None:
+        raise HTTPException(
+            status_code=403,
+            detail="Permission denied. Only admin users have access to this operation.",
+        )
     if movie_model is None:
         raise HTTPException(status_code=404, detail="Movie not found")
     db.delete(movie_model)
